services/alert-manager/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aio-pika==9.3.1
openai==1.84.0
python-dotenv==1.0.0 
azure-servicebus>=7.0.0
